everything works with search score and everything with hidden when no image is searched

+++ src/facial_recognization/Facial_recognizationDTO.java
@@ -11,7 +11,7 @@ public class Facial_recognizationDTO {
     public String email = "";
     public String image = "";
 	public boolean isDeleted = false;
-	public int match_parcentage = 0;
+	public double match_parcentage = 0;
 	
     @Override
 	public String toString() {
